requirements.txt
openai
langchain
python-dotenv
fastapi
uvicorn
streamlit
requests
